Refactor monolith system into separate systems
